Remove _utils.py
