--- scripts/_utils.py.lock
@@ -1,7 +1,0 @@
-version = 1
-revision = 3
-requires-python = ">=3.12"
-
-[options]
-exclude-newer = "0001-01-01T00:00:00Z" # This has no effect and is included for backwards compatibility when using relative exclude-newer values.
-exclude-newer-span = "P7D"
